innoshop/requirements.txt
Django==1.8.4
django-debug-toolbar==1.3.2
django-markitup==2.2.2
Markdown==2.6.2
pysqlite==2.8.0
sqlparse==0.1.16
wheel==0.24.0
